pyproject.toml
[tool.poetry]
name = "sky-segmentation"
version = "0.1.0"
description = ""
authors = ["haceneterbouche <hacene.terbouche@gmail.com>"]
readme = "README.md"
packages = [{include = "sky_segmentation"}]

[tool.poetry.dependencies]
python = "^3.8"
pytorch-lightning = "^1.7"
torch = "^1.12"
Pillow = "^9.2"
torchvision = "^0.13"
scikit-learn = "^1.1.3"
opencv-contrib-python = "^4.7.0.68"
opencv-python = "^4.6.0.66"
datasets = {extras = ["vision"], version = "^2.9.0"}
tqdm = "^4.64.1"
segmentation-models-pytorch = "^0.3.2"
mit-semseg = "^1.0.0"

[tool.poe.tasks]
force-cuda113 = "python -m pip install torch==1.12.1+cu113 torchvision==0.13.1+cu113 torchaudio==0.12.1+cu113 -f https://download.pytorch.org/whl/torch_stable.html"
force-cuda116 = "python -m pip install torch==1.12.1+cu116 torchvision==0.13.1+cu116 torchaudio==0.12.1+cu116 -f https://download.pytorch.org/whl/torch_stable.html"


[tool.poetry.group.dev.dependencies]
ipdb = "^0.13"
wandb = "^0.13.5"
matplotlib = "^3.6.2"
poethepoet = "^0.16.4"
ipywidgets = "^8.0.4"
jupyterlab = "^3.6.1"


[tool.poetry.group.lint.dependencies]
black = "^22.8"
flake8 = "^5.0"
isort = "^5.10"
mypy = ">=0.991"
pre-commit = "^2.16"
[tool.poetry.group.tests.dependencies]
pytest = "^7.1"
pytest-cov = "^3.0"
pytest-mock = "^3.8"
moto = "^4.0.5"


[tool.poetry.group.docs.dependencies]
Sphinx = ">=5.1"
myst-nb = "^0.16"
sphinx-autoapi = ">=1.8"
sphinx-rtd-theme = ">=1.0"

[tool.black]
target-version = ['py39']
line_length = 89
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
profile = "black"
line_length = 89
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true

[tool.semantic_release]
version_toml = "pyproject.toml:tool.poetry.version"
# to track and update version in other files:
# version_variable = ["some/file:__version__", ...]
branch = "main"
changelog_file = "CHANGELOG.md"
build_command = "pip install poetry && poetry build"
dist_path = "dist/"
upload_to_pypi = false
remove_dist = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
